# Minor cleanup of lz4 compressor