This results in a 5% speedup:

```
  old      html         654.1MB/s ±   13.5MB/s ( 2.07%) (N = 90, α = 99.9%)
  new      html         689.8MB/s ±   15.1MB/s ( 2.19%) (N = 90, α = 99.9%)
```